Interpolation Search

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -27,7 +27,7 @@ int interpolation_search(int *array, size_t size, int value)
 			(array[high] - array[low])) * (value - array[low]));
 
 		assert((low <= pos) && (pos <= high));
-		printf("Value checked array[%zu] = [%d]\n", pos, array[pos]);
+		printf("Value checked array[%lu] = [%d]\n", (unsigned long)pos, array[pos]);
 
 		if (array[pos] == value)
 			return (pos);
@@ -40,8 +40,8 @@ int interpolation_search(int *array, size_t size, int value)
 	pos = low + (((double)(high - low) /
 	(array[high] - array[low])) * (value - array[low]));
 	if (value < array[low])
-		printf("Value checked array[%zu] is out of range\n", pos);
+		printf("Value checked array[%lu] is out of range\n", (unsigned long)pos);
 	if (value > array[high])
-		printf("Value checked array[%zu] is out of range\n", pos);
+		printf("Value checked array[%lu] is out of range\n", (unsigned long)pos);
 	return (-1);
 }
